refactor(stages): move scan/metric progress wiring to internal/stages
