+++ cmd/codeviz/bubbletree_cmd.go
@@ -117,7 +117,7 @@ func (c *BubbletreeCmd) Run(flags *Flags) error {
 
 	slog.Info("Scanning filesystem", "path", c.TargetPath)
 
-	scanProg, stopScanTicker := buildScanProgress(flags)
+	scanProg, stopScanTicker := stages.BuildScanProgress(toStagesFlags(flags))
 
 	root, err := scan.Scan(c.TargetPath, filterRules, scanProg)
 
@@ -135,7 +135,7 @@ func (c *BubbletreeCmd) Run(flags *Flags) error {
 
 	slog.Info("Calculating metrics")
 
-	metricProg, stopMetricTicker := buildMetricProgress(flags, model.CountFiles(root))
+	metricProg, stopMetricTicker := stages.BuildMetricProgress(toStagesFlags(flags), model.CountFiles(root))
 
 	if err := provider.Run(root, requested, metricProg); err != nil {
 		stopMetricTicker()
